requirements.txt
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
openpyxl>=3.1.0
pyinstaller>=6.0.0
numpy>=1.24.0